+++ src/aegisx/scanners/__init__.py
@@ -7,4 +7,5 @@ __all__ = [
     "config_scanner",
     "dependency_scanner",
     "network_scanner",
+    "ssrf_scanner",
 ]